requirements.txt
pyyaml
django
